Removed the date from gemfile
Right now (according to rubygems.org) every release of hash_mapper was released on the same date. This seems to be because it's taking the date from the gemspec. I've removed that, so hopefully now the release date should be the same as the date it was published to rubygems.org